--- packages/cozy-dataproxy-lib/src/search/helpers/normalizeFile.spec.ts
@@ -1,286 +1,0 @@
-import CozyClient from 'cozy-client'
-import { IOCozyFile } from 'cozy-client/types/types'
-
-import {
-  addFilePaths,
-  computeFileFullpath,
-  normalizeFileWithFolders,
-  shouldKeepFile
-} from './normalizeFile'
-import { queryDocById } from '../queries'
-import { CozyDoc } from '../types'
-
-jest.mock('../queries', () => ({
-  queryDocById: jest.fn()
-}))
-
-const client = new CozyClient()
-
-describe('normalizeFileWithFolders', () => {
-  test('should get path for directories', () => {
-    const folders = [] as IOCozyFile[]
-
-    const file = {
-      type: 'directory',
-      dir_id: 'SOME_DIR_ID',
-      path: 'SOME/PATH/'
-    } as IOCozyFile
-
-    const result = normalizeFileWithFolders(folders, file)
-
-    expect(result).toStrictEqual({
-      type: 'directory',
-      _type: 'io.cozy.files',
-      dir_id: 'SOME_DIR_ID',
-      path: 'SOME/PATH/'
-    })
-  })
-
-  test('should get empty path for directories with missing path attribute', () => {
-    const folders = [] as IOCozyFile[]
-
-    const file = {
-      type: 'directory',
-      dir_id: 'SOME_DIR_ID'
-    } as IOCozyFile
-
-    const result = normalizeFileWithFolders(folders, file)
-
-    expect(result).toStrictEqual({
-      type: 'directory',
-      _type: 'io.cozy.files',
-      dir_id: 'SOME_DIR_ID',
-      path: ''
-    })
-  })
-
-  test(`should get parent folder's path for files`, () => {
-    const folders = [
-      {
-        type: 'directory',
-        _id: 'SOME_DIR_ID',
-        path: 'SOME/PARENT/PATH/'
-      }
-    ] as IOCozyFile[]
-
-    const file = {
-      type: 'file',
-      dir_id: 'SOME_DIR_ID'
-    } as IOCozyFile
-
-    const result = normalizeFileWithFolders(folders, file)
-
-    expect(result).toStrictEqual({
-      type: 'file',
-      _type: 'io.cozy.files',
-      dir_id: 'SOME_DIR_ID',
-      path: 'SOME/PARENT/PATH/'
-    })
-  })
-
-  test(`should get empty path for files when parent forlder is missing path attribute`, () => {
-    const folders = [
-      {
-        type: 'directory',
-        _id: 'SOME_DIR_ID'
-      }
-    ] as IOCozyFile[]
-
-    const file = {
-      type: 'file',
-      dir_id: 'SOME_DIR_ID'
-    } as IOCozyFile
-
-    const result = normalizeFileWithFolders(folders, file)
-
-    expect(result).toStrictEqual({
-      type: 'file',
-      _type: 'io.cozy.files',
-      dir_id: 'SOME_DIR_ID',
-      path: ''
-    })
-  })
-})
-
-describe('addFilePaths', () => {
-  it(`should add parent folder's path to files`, () => {
-    const docs = [
-      {
-        _id: 'SOME_FILE_ID',
-        _type: 'io.cozy.files',
-        type: 'file',
-        dir_id: 'SOME_DIR_ID',
-        name: 'myfile.txt'
-      },
-      {
-        _id: 'SOME_DIR_ID',
-        _type: 'io.cozy.files',
-        type: 'directory',
-        path: '/mydir',
-        dir_id: 'ROOT_ID'
-      }
-    ] as CozyDoc[]
-
-    const result = addFilePaths(docs)
-
-    expect(result[0]).toStrictEqual({
-      _id: 'SOME_FILE_ID',
-      _type: 'io.cozy.files',
-      type: 'file',
-      dir_id: 'SOME_DIR_ID',
-      name: 'myfile.txt',
-      path: '/mydir/myfile.txt'
-    })
-  })
-
-  it(`should handle no files in results`, () => {
-    const result = addFilePaths([])
-    expect(result).toStrictEqual([])
-  })
-
-  it(`should handle when no parent dir is found`, () => {
-    const docs = [
-      {
-        _id: 'SOME_FILE_ID',
-        _type: 'io.cozy.files',
-        type: 'file',
-        dir_id: 'SOME_DIR_ID'
-      }
-    ] as CozyDoc[]
-
-    const result = addFilePaths(docs)
-
-    expect(result).toStrictEqual([
-      {
-        _id: 'SOME_FILE_ID',
-        _type: 'io.cozy.files',
-        type: 'file',
-        dir_id: 'SOME_DIR_ID',
-        path: ''
-      }
-    ])
-  })
-})
-
-describe('shouldKeepFile', () => {
-  test(`should return true for expected files`, () => {
-    const file = {
-      _id: 'SOME_FILE_ID',
-      _type: 'io.cozy.files',
-      type: 'file',
-      dir_id: 'SOME_DIR_ID'
-    } as unknown as IOCozyFile
-
-    const result = shouldKeepFile(file)
-
-    expect(result).toBe(true)
-  })
-
-  test(`should return false for trashed files`, () => {
-    const file = {
-      _id: 'SOME_FILE_ID',
-      _type: 'io.cozy.files',
-      type: 'file',
-      dir_id: 'SOME_DIR_ID',
-      trashed: true
-    } as unknown as IOCozyFile
-
-    const result = shouldKeepFile(file)
-
-    expect(result).toBe(false)
-  })
-
-  test(`should return false for files in trash`, () => {
-    const file = {
-      _id: 'SOME_DIR_ID',
-      _type: 'io.cozy.files',
-      type: 'file',
-      path: '/.cozy_trash/SOME_FILE'
-    } as unknown as IOCozyFile
-
-    const result = shouldKeepFile(file)
-
-    expect(result).toBe(false)
-  })
-
-  test(`should return false for root dir`, () => {
-    const file = {
-      _id: 'io.cozy.files.root-dir',
-      _type: 'io.cozy.files',
-      type: 'directory'
-    } as unknown as IOCozyFile
-
-    const result = shouldKeepFile(file)
-
-    expect(result).toBe(false)
-  })
-
-  test(`should return false for shared drives`, () => {
-    const file = {
-      _id: 'io.cozy.files.shared-drives-dir',
-      _type: 'io.cozy.files',
-      type: 'directory'
-    } as unknown as IOCozyFile
-
-    const result = shouldKeepFile(file)
-
-    expect(result).toBe(false)
-  })
-})
-
-describe('computeFileFullpath', () => {
-  const dir = {
-    _id: '123',
-    _type: 'io.cozy.files',
-    type: 'directory',
-    dir_id: 'ROOT',
-    name: 'MYDIR',
-    path: 'ROOT/MYDIR'
-  } as IOCozyFile
-  const fileWithFullpath = {
-    _id: '456',
-    _type: 'io.cozy.files',
-    type: 'file',
-    dir_id: '123',
-    name: 'file1',
-    path: 'ROOT/MYDIR/file1'
-  } as IOCozyFile
-  const fileWithStackPath = {
-    _id: '789',
-    _type: 'io.cozy.files',
-    type: 'file',
-    dir_id: '123',
-    name: 'file2',
-    path: 'ROOT/MYDIR'
-  } as IOCozyFile
-  const filewithNoPath = {
-    _id: '000',
-    _type: 'io.cozy.files',
-    type: 'file',
-    dir_id: '123',
-    name: 'file3'
-  } as IOCozyFile
-  it('should handle directory', async () => {
-    const res = await computeFileFullpath(client, dir)
-    expect(res).toEqual(dir)
-  })
-
-  it('should handle file with complete path', async () => {
-    const res = await computeFileFullpath(client, fileWithFullpath)
-    expect(res).toEqual(fileWithFullpath)
-  })
-
-  it('should compute fullpath for file with incomplete path', async () => {
-    const res = await computeFileFullpath(client, fileWithStackPath)
-    expect(res.path).toEqual('ROOT/MYDIR/file2')
-  })
-
-  it('should compute fullpath for file with no path', async () => {
-    // eslint-disable-next-line prettier/prettier
-    (queryDocById as jest.Mock).mockResolvedValue(dir)
-
-    const res = await computeFileFullpath(client, filewithNoPath)
-
-    expect(res.path).toEqual('ROOT/MYDIR/file3')
-  })
-})
